+++ CHANGELOG.md
@@ -6,17 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2021-10-18
+### Added
+- Save EEG raw data to csv file
+- Raw data chart add pause function to halt data update
+- Choice box which can select connection COM port
+
 ## [0.3.0] - 2021-10-08
 ### Added
 - EEG raw data viewer
-- event record what time receive it
+- Event record what time receive it
 ### Changed
-- move all file to com.flyn.eeg_receiver folder
+- Move all file to com.flyn.eeg_receiver folder
 - MindSetReceiver class renamed EEGReceiver class
 
 ## [0.2.2] - 2021-10-01
 ### Fixed
-- dataDecode throw the null pointer exception
+- DataDecode throw the null pointer exception
 
 ## [0.2.1] - 2021-09-27
 ### Added